+++ examples/list_ipsec_and_virtual_circuits_in_tenancy/list_all_virtual_circuits_in_tenancy.py
@@ -45,6 +45,7 @@
 #   -t config - Config file section to use (tenancy profile)
 #   -p proxy  - Set Proxy (i.e. www-proxy-server.com:80)
 #   -ip       - Use Instance Principals for Authentication
+#   -dt       - Use Instance Principals with delegation token for cloud shell
 ##########################################################################
 from __future__ import print_function
 import sys
@@ -52,6 +53,7 @@ import argparse
 import datetime
 import oci
 import json
+import os
 
 
 ##########################################################################
@@ -81,10 +83,10 @@ def check_service_error(code):
 
 ##########################################################################
 # Create signer for Authentication
-# Input - config_profile and is_instance_principals
+# Input - config_profile and is_instance_principals and is_delegation_token
 # Output - config and signer objects
 ##########################################################################
-def create_signer(config_profile, is_instance_principals):
+def create_signer(config_profile, is_instance_principals, is_delegation_token):
 
     # if instance principals authentications
     if is_instance_principals:
@@ -97,8 +99,55 @@ def create_signer(config_profile, is_instance_principals):
             print_header("Error obtaining instance principals certificate, aborting")
             raise SystemExit
 
+    # -----------------------------
+    # Delegation Token
+    # -----------------------------
+    elif is_delegation_token:
+
+        try:
+            # check if env variables OCI_CONFIG_FILE, OCI_CONFIG_PROFILE exist and use them
+            env_config_file = os.environ.get('OCI_CONFIG_FILE')
+            env_config_section = os.environ.get('OCI_CONFIG_PROFILE')
+
+            # check if file exist
+            if env_config_file is None or env_config_section is None:
+                print("*** OCI_CONFIG_FILE and OCI_CONFIG_PROFILE env variables not found, abort. ***")
+                print("")
+                raise SystemExit
+
+            # check if file exist
+            if not os.path.isfile(env_config_file):
+                print("*** Config File " + env_config_file + " does not exist, Abort. ***")
+                print("")
+                raise SystemExit
+
+            config = oci.config.from_file(env_config_file, env_config_section)
+            delegation_token_location = config["delegation_token_file"]
+
+            with open(delegation_token_location, 'r') as delegation_token_file:
+                delegation_token = delegation_token_file.read().strip()
+                # get signer from delegation token
+                signer = oci.auth.signers.InstancePrincipalsDelegationTokenSigner(delegation_token=delegation_token)
+
+                return config, signer
+
+        except KeyError:
+            print("* Key Error obtaining delegation_token_file")
+            raise SystemExit
+
+        except Exception:
+            raise
+
+    # -----------------------------
     # config file authentication
+    # -----------------------------
     else:
+        # check if file exist
+        if not os.path.isfile(oci.config.DEFAULT_LOCATION):
+            print("*** Config File " + oci.config.DEFAULT_LOCATION + " does not exist, Abort. ***")
+            print("")
+            raise SystemExit
+
         config = oci.config.from_file(
             oci.config.DEFAULT_LOCATION,
             (config_profile if config_profile else oci.config.DEFAULT_PROFILE)
@@ -146,16 +195,17 @@ parser = argparse.ArgumentParser()
 parser.add_argument('-t', default="", dest='config_profile', help='Config file section to use (tenancy profile)')
 parser.add_argument('-p', default="", dest='proxy', help='Set Proxy (i.e. www-proxy-server.com:80) ')
 parser.add_argument('-ip', action='store_true', default=False, dest='is_instance_principals', help='Use Instance Principals for Authentication')
+parser.add_argument('-dt', action='store_true', default=False, dest='is_delegation_token', help='Use Delegation Token for Authentication')
 cmd = parser.parse_args()
 
 # Start print time info
 start_time = str(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
-print_header("Running List IPSecTunnels")
+print_header("Running List List Virtual Circuits")
 print("Starts at " + str(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")))
 print("Command Line : " + ' '.join(x for x in sys.argv[1:]))
 
 # Identity extract compartments
-config, signer = create_signer(cmd.config_profile, cmd.is_instance_principals)
+config, signer = create_signer(cmd.config_profile, cmd.is_instance_principals, cmd.is_delegation_token)
 compartments = []
 tenancy = None
 try: